refactor(kwin-effect): extract window-id helpers to plasmazoneseffect/window_identity.cpp
Pulls getWindowId, getWindowInstanceId, getWindowAppId, pushWindowMetadata,
flushPendingFrameGeometry, isPlasmaShellSurface out of plasmazoneseffect.cpp
into a new TU. Drops the main file from 5257 to 5138 lines.

No API or behavior change.
